Fixed fadewave halting after one wave and removed obsolete update(BRIGHTNESS_TYPE) methods.
Removed the check for remaining waves to fade back up after a fade down.
This check made the fadewaving stop, because if the led is fading continuously, the number of remaining fadewaves is zero. The function "isFadeWaving" returns true if the latter is the case, or when the remaining fadewaves are not zero.

Also removed obsolete update(BRIGHTNESS_TYPE) methods.